fix: address top React Doctor findings
