+++ frontend/package.json
@@ -41,6 +41,6 @@
     "typescript-eslint": "^8.58.2",
     "vite": "^8.0.10",
     "vite-imagetools": "^10.0.0",
-    "vitest": "^3.0.5"
+    "vitest": "^4.1.9"
   }
 }
