+++ frontend/src/components/ui/ProgressBar.tsx
@@ -11,8 +11,8 @@
 export type ProgressBarVariant = 'progress' | 'learning';
 
 const fillClasses: Record<ProgressBarVariant, string> = {
-  progress: 'bg-salem',
-  learning: 'bg-learning',
+  progress: '[&::-webkit-progress-value]:bg-salem [&::-moz-progress-bar]:bg-salem',
+  learning: '[&::-webkit-progress-value]:bg-learning [&::-moz-progress-bar]:bg-learning',
 };
 
 export function ProgressBar({
@@ -26,16 +26,11 @@ export function ProgressBar({
 }) {
   const clamped = Math.min(100, Math.max(0, value));
   return (
-    <div className="h-1 bg-surface-elevated rounded-full overflow-hidden">
-      <div
-        className={`h-full rounded-full ${fillClasses[variant]}`}
-        style={{ width: `${clamped}%` }}
-        role="progressbar"
-        aria-label={label ?? 'Course progress'}
-        aria-valuenow={clamped}
-        aria-valuemin={0}
-        aria-valuemax={100}
-      />
-    </div>
+    <progress
+      className={`h-1 w-full appearance-none overflow-hidden rounded-full bg-surface-elevated [&::-webkit-progress-bar]:rounded-full [&::-webkit-progress-bar]:bg-surface-elevated [&::-webkit-progress-value]:rounded-full [&::-moz-progress-bar]:rounded-full ${fillClasses[variant]}`}
+      value={clamped}
+      max={100}
+      aria-label={label ?? 'Course progress'}
+    />
   );
 }
